+++ src/demo-app/app/layout/layout.menu.ts
@@ -23,7 +23,7 @@ export const OPTIONS: MenuOption[] = [
    { route: 'alerts-demo', name: 'Alerts' },
    { route: 'dropdown-menu-demo', name: 'Dropdown Menu' },
    { route: 'dropdown-demo', name: 'Dropdown' },
-   { route: 'combobox-demo', name: 'Combobox' },
+   { route: 'select-demo', name: 'Select' },
    { route: 'help-demo', name: 'help' },
    { route: 'info-box-demo', name: 'info-box' },
    { route: 'item-list-demo', name: 'item-list' },